+++ project/main/AlgorithmFlow.h
@@ -17,7 +17,7 @@ class AlgorithmFlow {
     // Each point on GroundTruthFlow is a vector of points in AlgorithmFlow. Hence both the base and fast movement
     // consists of an additional vector wrappper.
 
-    std::vector<std::vector<std::vector<std::pair<cv::Point2i, cv::Point2i> > > > m_algo_frame_pixel_point_pixel_displacement;
+    std::vector<std::vector<std::vector<std::pair<cv::Point2f, cv::Point2f> > > > m_algo_frame_pixel_point_pixel_displacement;
     std::string m_resultordner;
 
 public:
@@ -27,10 +27,10 @@ public:
     void calculate_flow(ALGO_TYPES algo, FRAME_TYPES frame_types, NOISE_TYPES noise);
 
     void storeData(const std::vector<cv::Point2f> &prev_pts, std::vector<cv::Point2f> &next_pts, const
-    std::vector<uchar> status, std::vector<std::vector<std::pair<cv::Point2i,
-            cv::Point2i> > > &frame_pixel_point_pixel_displacement);
+    std::vector<uchar> status, std::vector<std::vector<std::pair<cv::Point2f,
+            cv::Point2f> > > &frame_pixel_point_pixel_displacement);
 
-    void store_in_yaml(cv::FileStorage &fs, const cv::Point2i &l_pixelposition, const cv::Point2i
+    void store_in_yaml(cv::FileStorage &fs, const cv::Point2f &l_pixelposition, const cv::Point2f
     &l_pixelmovement );
 
     void setResultOrdner(ALGO_TYPES algo, FRAME_TYPES frame_types, NOISE_TYPES noise);
